Generates Nightwatch configuration dynamically. Closes #3.
It allows for an override file to input project specific information
in the test module without messing with the defaults. This override JSON
can include further information about the project (including a path to
another JSON, such as a package.json) which will be used to populate
further metadata about the project.